Fixing bugs in reading, for pftool and fuse.
-- there were bugs in reading, resulting from the fact that the response
header appears to arrive long before the data is completed.
stream_writeheaderfunc() parses out the result-code (e.g. 206 "Partial
Content").  But stream_get(), and stream_open() which it sometimes calls,
were calling aws_iobuf_reset(), which cleared the result-code.  This meant
that, upon completion of the request, there was no result-code.  Replaced
aws_iobuf_reset() calls with calls to the new aws_iobuf_reset_lite(), which
just resets buffers, but not the codes.

NOTE: This implies that the s3_op thread receiving writeheaderfunc
callbacks, and the user-interaction thread calling stream_get(), are making
unprotected asynchronous accesses to the same IOBuf struct.  However, they
are accessing different fields.  The writeheaderfunc only installs
header-fields, which the user-threads avoid touching.  By the time
stream_wait() reutrns (after the s3_op), all parallel access are complete.

-- Reworked marfs_read() quite a bit.

-- Adapted some of the technology used to manage writes (in
MarFS_FileHandle.WriteStatus) to similar support in ReadStatus.  Added
byte-ranges to all curl read-requests.  In some cases, this can reduce the
amount of data moved.  For example, reading from a PACKED object, or
reading the tail of an N:1 segment which ends in the middle of an object.
These need only move the data they actually will use.  And, in all cases,
we avoid moving the recovery-info at the tail of objects.

For fuse, we can only limit the ranges to the simplest cases (e.g. a PACKED
object, one object within a MULTI, or the full extent of a UNI).  At least
we leave out reading the recovery-info.  For pftool, we do know the total
read-length, and we can mange it across objects in a multi, when reading an
N:1 segment that spans multiple objects.

Ultimately, for both fuse and pftool (but especially for pftool), our
tracking of the total expected range of the read (across MULTI objects)
could allow us to overlap the overhead of opening the next object on top of
reads in the current object, because we actually know whether we will need
it or not.

-- Changed FH_ALLOW_RISKY to FH_Nto1_WRITES, which seems more meaningful.

Stripped memsets out of pftool, for speed.  Valgrind shows a lot of time
spent doing this, but it's a waste: if the reads succed, they'll overwrite
the buffer with new contents anyhow.  If the reads don't succeed, it's
considered a failure.  It appears that there is no danger of uninitialized
buffer-contents getting into files.