Refactored fns managing PRE xattr, so Alfred can share.  Bug fix in reads.
The somewhat-recent addition of host-randomization was handled inline in
marfs_read().  Moved that to init_pre() [generating the random-number
seed], and update_pre() [computing a random IP-address within a given
range.  This should make this technique accessible in the GC/quota scripts.
(They don't need randomization, but the host ip-address provided in
configurations that want to use randomization are not usable until one
performs that randomization.)

This turned up a bug in the recent changes to read, where the chunk-number
in the object-ID was not getting updated.  Fixed.

Also tweaked/extended the functions that read chunkinfo from the MD file.
We aren't currently using these, but it has come up in discussions of the
efficiency of restart in pftool.