seed for host-randomization has moved from PathInfo to Pre
This allows it to be accessible to tools that are not tightly integrated
with MarFS I/O, such as the quota/GC tools, which want to do minimal
parsing of xattrs, and need to have access to a host IP-address, for
constructing full URLs.  This code is now not required to deal at the level
of PathInfo, to get that done using library support.