Previous re-factoring of host-randomization worked for writes, but not reads.
str_2_pre(), instead of init_pre(), is used in the case of parsing existing
PRE xattr values.  In this case, we weren't computing an initial
random-seed, so the first call to update_pre(), to generate a host was
always getting the same initial-value, meaning that all radomizations were
getting the same host.