Can use hostname-hash, instead of random-in-range, to find server IP.
We've been digging into an incast problem for parallel reads.  This crops
up when pftool is running multiple read tasks on a given host, using the
old host-randomization, where every request picks the server at random from
a range of addresses.  The old approach was a quick-and-dirty way to
provide fail-over.  If a host stops responding: try a different one.

The problem is that multiple tasks on the same host, picking server IPs at
random, can lead to a "TCP throughput collapse" when the requests collide
at the client switch port, and packets are dropped.

Some recent testing indicates that we can address this by breaking large
GET requests up into a sequence of smaller requests (i.e. for sub-ranges).
This is less likely to overflow the buffers on the recv port.  This also
has the advantage of allowing host-randomization to work again, which
provides cheap, local fail-over.  But we haven't implemented that, yet.
(Our problem may also go away on the faster networks we'll see on Trinity.)

Meanwhile, we can also address it by having all the tasks on a given host
make their requests to the same server.  So, requests pick their server
from the range using a hash of the local host name.  That way, the recv
switch-port doesn't have to buffer more than the sending port can transmit,
which is presumably the same.  So, that's what we're doing here.  This
should resolve some of our incast issues, at the expense of giving up
fail-over, for now.