Add repo.max_request_size.  More-fully-assed versioning.
A new Repo-parameter max_request_size can cause larger reads (GETs) to be
transparently broken up into a series of smaller requests.  This was
introduced during our efforts to respond to an apparent 'incast' problem on
ODSU FTAs, where it seemed to help.  May be obviated by network tuning
we've subsequently done.

Seemed appropriate to bump the "config file version" for the first time
ever.  max_request_size in earlier versions of configuration will default
to 0, which means there is no additional constraint on request-size, and
read requests will just be complained by chunk-sizes, etc, as before.

This issue with defaulting earlier config files lead to the realization
that the "config version" was poorly defined.  The real intent was to
identify changes in xattrs, or object-layout, or MD file layout.  But it
also pertains to changes in the config-file per se.  The two are related
but changes in either one may or may not correspond with changes in the
other.  This is discussed in marfs_configuration.h, and in
marfs_configuration_blueprint.cfg.  Bottom line is that, for now, we will
understand "marfs version" as the #defines that are in
marfs_configuration.h.  This is what will be written into object IDs, and
it should change whenever there are changes in any of: config-file,
config-reader, object-ID, recovery-info, MD layouts, etc.