Fleshed out marfs_statvfs().  Root NS.md_path points to MDFS.
[NOTE: This was formerly known as marfs_statfs()]

marfs_statvfs() was just returning "not supported", but this is now needed
to support doing an NFS export of a marfs fuse mount.

If you install the mount-point of the MDFS as the md_path in the root
namespace of your configuration, statvfs() of the marfs fuse mount will now
return the stavfs of the MDFS.  If you don't do that, or if the internal
statvfs() fails, we return defaults suggesting the maximum possible
available space.

Updated root NS in marfs_configuration.cfg and marfs_odsu.cfg, and added
comments to relevant parts of the blueprint file.

TBD: Use the fsinfo files to gather the actual amount of used space.  We
are prepared for an optimization of that approach, where the tool that
crawls GPFS to count used storage in each namespace (in support of quota
enforcement), could also store the sum of all those counts into an fsinfo
for the root NS.