marfs_statvfs() will use individual namespaces, if needed. Fixed defaulting.
Previosuly we tried to always use the root_ns, or fail-over to defaults.
That wasn't working right for the backward-compatibility it was intended to
support.  Now, old config files should default properly, for the root
namespace, in the case where we used a non-existent path for the md_path.
That is now the only case that defaults.  For everything else, we call
statvfs() on the configured md_path, and return whatever that returns.
This will perform real permission-checking, as well.