Deny access to incompletely-written files.  Fix a bug in close().
We install the marfs_restart xattr when files are opened, and remove it
when files are closed.  In the case of N-to-1 files, pftool may have
multiple concurrent tasks doing the writing, so close() [aka
marfs_release()] can't remove the restart xattr, in that case.  Instead,
batch_post_process() does that.

Now that we've straightened this all out, marfs_restart indicates a file
that was not completely written (e.g. maybe it's N-to-1, and some of the
writer-tasks failed).  pftool can be rerun to fix such problems (e.g. to
write the missing parts).  Meanwhile, MarFS internals will return EINVAL,
if you attempt to read such files.

However, we were not correctly maintaining marfs_restart in the past, so
files that were correctly written will still have this xattr, and reads of
such old MarFS files will fail, after installing this patch.  If you are
willing to assume that they are all actually correct, you could do
something like the following to make them readable again:

    #!/bin/bash
    # /gpfs/marfs-gpfs/jti/mdfs is the namespace.md_path from the config.
    # Replace with the appropriate path from your config
    for F in `find /gpfs/marfs-gpfs/jti/mdfs -type f | sort`; do
       echo $F
       attr -r marfs_restart $F 2>&1
    done

In addition, there was a bug in marfs_release() that may have been causing
close-failures, on writes to MarFS, for some configurations.  Fixed.