Replace process group-list with user group-list, for some fuse ops.
This solves serveral issues we've been having with odd behavior after
seteuid()/setegid().  It turns out Linux processes have their own
group-lists, and the group-related access-controls on files use this
group-list.  setgid()/setegid()/setfsgid() merely add an item to the
per-process group-list.  setuid()/seteuid()/setfsuid() don't affect it at
all.

So, if we run fuse as root, and expect any of these functions to give us
the temporary appearance of running as a given user/group, we get
surprising results.  They can see into directories they shouldn't be able
to, and can't chgrp files to a group that they are a member of, if it isn't
their default group.

Those problems are fixed now.  More discussion in the github issues.