cleanup curl handles at the tail of stream_sync(), to reduce CLOSE_WAITs
Reusing curl-handles (which we don't do yet) is expected to buy us some
performance improvement.  But, for now, the goal was to reduce the number
of TCP connections left in CLOSE_WAIT.  Each of our stream threads
currently gets its own curl-handle.  When we close the stream, we now also
cleanup the handle.  That, at least, prevents the connetion being left in
CLOSE_WAIT.