Touched up recovery-info section in Design-Config document.
I don't believe I've changed things in a deep way, but added/tweaked some
details.  For example, it seems we need to store the obj-ID itself, because
sproxyd doesn't keep track of the object-IDs we used to write objects.  We
can recover most of what what is needed to recreate the obj-ID, but we
might as well just skip that and store the object-ID.

Also, storing the data-length, rather than data-offset, allows the
RecoveryInfo records to be "relocated" without alteration, during packing
by a post-processing script.