Writing real recovery-info into the tails of all objects.
The intent is that all objects now carry enough recovery-info to allow us
to regenerate the meta-data for the associated MarFS files, should there be
a meltdown in the MDFS.

The current format should also allow relatively painless packing of
objects, in which data and recovery-info for multiple small objects is
combined into a single larger one.  One would throw out the "tail" portion
of the rec-info for the former objects, and write a new "tail" portion at
the end of teh combined object, but the rest of the rec-info can be
relocated as needed without changes.