New URLs, for rev 1.0.  http://host:port/ns_alias/repo/version/ns/...
This is incompatible with previous releases.  All object-IDs written with
previous releases will not be readable with this one.  (Object-IDs do have
version numbers, so it would be possible to add version-specific behavior
to parsers to allow old and new objects to co-exist, but we didn't do that.
We're wiping our old objects, to prepare for the new release.)

There are updates to marfs_cctest.cfg and marfs_odsu.cfg for the new setup.

The new scheme has several advantages.

(a) You can create a new namespace without any modifications to server
config.  For Scality sproxyd, the new "namespace alias" field in the URL,
corresponds to a fastcgi proxy name.  There is a corresponding field in the
configuration.  Many namespaces can use the same proxy, so no need to add a
new one to the server, just to add a new namespace.  In the case of S3,
this alias corresponds with a bucket.  Thus, you can give each namespace
its own bucket, or merge them into a common place.

(b) repo now sits in the second token, instead of the first.  For sproxyd,
this corresponds to an sproxyd driver-alias.  Because the driver-alias can
have a unique erasure-coding schema, and other "repo-like" features, this
makes sense as a way to allow creating MarFS "repos" in the same Scality
ring.