Added per-repo config-settings for read/write timeouts.
This lets admins tweak the timeouts used for read/write, without having to
get changes into the code.  You can add either or both of the specs, or
leave them out.  They default to 0, which will ultimitely defer to the old
hardwired values in the code.  [The hardwired defaults are currently 20
seconds, for both reads and writes.]

    # example ...
    <repo>
      ...
      <write_timeout>20</write_timeout>
      <read_timeout>10</read_timeout>
    </repo>

TBD: Move the hardwired values out of object_stream.c to some .h file, so
they can be manipulated more easily.

TBD: stream_wait() actually has its own default_timeout (currently also 20
seconds).  For now, it uses the same timeout as was set for the
corresponding get/put.  We could make this a separate config, if we want.