Added MARFS_MAX_NAMESPACE_NAME back into #defines.
Not needed anymore, now that Namespace.name is no longer put into URLs in a
place corresponding to the S3 "bucket" (which has constrained length).
However, some clients would like to know how big to allocate static buffers
that receive these names.  Well, we know they can't be bigger than an
obj-ID, so we amke that the new #defn.