Introducing MDAL (unused).  Fixed compiler warnings.
The MetaData Abstraction Layer (MDAL) is going to allow us to experiment
with alternative implementations of the MD file-system.  This will come in
handy for testing the scalability of sharded MD directories, when we get to
that.  For now, this is just a modular check-in, so others (who are
planning on implementing an MDFS via OrangeFS) can see the framework.

We are not yet actually using any of the MDAL in libmarfs.

There is also some reworking of Makefile (adding universal
-D_XOPEN_SOURCE=700) and some .h's, to clean up compiler warnings.  Quick
test shows that fuse seems to work fine, and pftool builds.