MDAL (still unused) now integrated with configuration.
Each namespace can specify dir_MDAL and file_MDAL.  The spec is one of
POSIX, PVFS2, or IOFSL.  Only POSIX is implemented; the others will log an
error and the initialization will abort.  This is backward-compatible.  If
either spec is missing from the config file, it will default to POSIX.

At read-config time, MDALs matching the specified (legal) types for
dir_MDAL and file_MDAL are initialized and installed into the NS.

TBD: Replace the md_fd in MarFS_FileHandle, and the DIR* in
MarFS_DirHandle, with MDAL_Contexts.  Make appropriate initializations.
Replace all MDFS accesses with MDAL function-calls.