Changed rec-info tail to have rec-info len, rather than offset, for packer.
Incremented the global format-version from 1.0 to 1.1, so the packer can
recognize those objects with this new property.  It is only these newer
objects that are amenable to recovery, if they are packed in the "data1,
rec1, tail1, data2, rec2, tail2, ..." style, rather than the "data1, data2,
... rec1, rec2, ..., tail" style.  This style (the first mentioned) is
more-efficient for the packer, though maybe more awkward for recovery.
However, it seems to make more sense to optimize for packing.