Interagrated MDAL with libmarfs.
For now, the MDAL interaction is conditional; you must build libconfig and
libmarfs with USE_MDAL=1, to select MDAL interaction.  This is assimilated
into the appropraite makefiles:

  (cd $CONFIG/src; make clean && make USE_MDAL=1)
  (cd $FUSE/src;   make clean && make USE_MDAL=1)

Add DEBUG=x as appropriate for each build, if you want to debug.  You can
do this comprehensively with the build_marfs script:

  $FUSE/scripts/build_marfs 2 1

Either way, Namespace.file_MDAL and Namespace.dir_MDAL (in the MarFS
configuration file) will be used to define the interface for file-oriented
operations, and directory-oriented operations, respectively.