Improved handling of chunk-info MD files for N:1
Don't open the MD files by default in marfs_open/write, for Nto1.  This
allows pftool to do all the MD updates from the update-task (rank 2), so
updates to N:1 chunk-info MD are all done from a single task.  This should
improve performance on GPFS (and allow the system to work at all, on
non-parallel FS).