Fixed issue where we were writing recovery-info even when ftruncate failed.
This applies to the case where a user is overwriting a file, but there is a
misconfigured trash-directory, such that the user doesn't have access.  We
would err out of truncating the file, but would leave the object stream
open.  release() would then write recovery-info into the object, and update
xattrs, so the file became a legitimate empty file.  After the fix, the
obj-stream is aborted earlier in ftruncate(), so release() can detect that
and skip out of doing the updates.