Changes to support pftool N:1 restart where CTM is behind by serveral objects.
In pftool, file-CTM is only updated periodically, so it may not have heard
of all the chunks that have been written to a destination.  Therefore,
during pftool restart for an N:1 copy to marfs, some of these destination
chunks may be re-written.  However, Scality sproxyd doesn't allow object
overwrites (for performance reasons).  Instead, we get a 500, and we only
get it at the tail of the chunk, in the stream_sync() after writing
recorvery-info, in marfs_write().

pftool will detect and handle this situation, as a failed call to
marfs_write().  marfs_release() should then be able to call stream_sync()
and stream_close() without bailing out on the errors they will report.

We add a new OSF_THREAD_ERR flag, which is convenient for marfs_release().