Added per-file-handle lock to read(), for exporting marfs fuse via NFS.
This makes file-handles thread-safe for reading.  (Similar approach could
be applied to write(), but we are currently only intending to export fuse
read-only.)

When a fuse mount is exported over NFS, nfsd sometimes makes concurrent
calls to read() at different offsets, on the same file-handle.  Without
this patch, such reads would attempt to operate the same object-stream on
behalf of different readers, managing from the same info in the
file-handle, leading to curl errors, or worse.  This fixes that.

However, we also had an issue with how discontiguous reads were handled.
We want to allow discontiguous reads for a reader that is genuinely
seeking.  But that requires closing the stream and reopening at a new
offset.  If we do that on every detected discontiguous read in the case of
NFS, the NFS performance is abysmal.  The current approach makes a crude
compromise: In the event we discover a discontiguous read, we realease the
read lock, wait 20 ms, and try to acquire again.  Up to 5 times.  This
allows concurrent readers the chance to get in, read through the stream up
to the spot of our discontiguous request, and allow us to proceed, without
a close/reopen.  If the gap is too large, we will close and reopen, then
one of the earlier reads will show up, and we'll close and reopen again.
But it's better than flailing on every one of them.

The addition of locking doesn't seem to noticibly affect performance of a
single-threaded read through fuse, and the retry business is allowing us to
get about half of the normal fuse BW over NFS (because of the cost of
numerous 20ms stalls).

TBD: A better scheme for handling discontiguous reads would be to queue the
waiters in order of offset, and let them retry immediately when anyone
releases the read-lock.