Stupid NFS tricks, for exporting fuse on top of object-storage repos.
nfsd calls fuse with multiple threads reading at diffrerent offsets on the
same file-handle.  The generic marfs method for dealing with such
out-of-order offsets is to close the GET-stream, and re-open at the new
offset.  This makes for miserable performance through NFS.

The new approach is to detect this situation and start enqueueing those
reads that specify an offset beyond the current stream-offset.  We do the
detection by stalling out-of-order reads and noticing whether the stream
changes.  An enqueued read-thread is waiting on a lock.  If other ongoing
reads bring the offset for an enqueued read into place, then that enqueued
read is dequeued and unblocked, and can proceed.  A timeout allows the
head-of-queue to eventually force the close/reopen, if needed.