Store overly-restrictive mode-bits in RESTART xattr, while a file is open.
Fuse calls mknod() with the desired final mode-bits.  If these are RO or
WO, we'll have problems manipulating xattrs in release().  However, mknod()
just gets a path, not a file-handle.  We were already storing a RESTART
xattr in mknod() to communicate with open() [etc], that this file is not
DIRECT.  Extend that concept by putting the final mode-bits into the
RESTART xattr (in this case only), and assure that user-write and user-read
are asserted.  Then (in this case only) release() can install the
originally-intended mode-bits.