Makefile now builds differently-named libraries for MDAL/non-MDAL.
libmarfs_mdal/libmarfs_nomdal.  We always provide a libmarfs symlink, for
apps that don't care.  Not sure if this is a great idea.  Started out as an
approach to pftool issue #31, allowing us to return to forcing filehandles
to only support MDAL or non-MDAL (i.e. getting the compiler to prevent
app-developers from hurting themselves).  But that would require all app
builds to reliably define USE_MDAL at compile-time, consistently with the
library they are linking, which is still error-prone.  Therefore, we still
keep both fd and MDAL-support in file-handles, and rely on app-developers
to never ever ever screw around with file-handle guts.