Initial implementation of support for a Data Abstraction Layer (DAL)
The MDAL applies to meta-data manipulation, which has POSIX semantics.
The DAL applies to storage manipulation, and has REST(-like) semantics.

This first cut includes configuration support added to the
configuration-options for Repos.  You can add <DAL>dal_type</DAL> into the
config for a repo, where "dal_type" can be one of: (a) OBJ [meant to apply
to our legacy object-storage implementation], (b) MD_DEMO, (C) MC, or (d)
POSIX [to allow a simple storage technique that could work "on a laptop",
etc].  If nothing is specified, repos default to DAL type = OBJ.

The build_marfs script now takes a third positional arg, to enable/disable
building to use the DAL, though this doesn't yet engage any code.  The idea
is that we will first compare the OBJ implementation with the legacy
data-access via object_stream.  In this case, the DAL is just an
indirection that ultimately calls the same functions.

TBD: Finish macros (like DAL_OPEN, etc) in dal.h that can be used for all
  existing calls to storage-related functions, so that the USE_DAL
  build-option will cause us to use the DAL or not.  These should replace
  all the calls that currently use the inteface in object_stream.h.

TBD: Do we need to rethink MarFS_FileHandle?  It looks like everything in
  there that isn't MDAL related is DAL-related.  That's going to be a
  complex mess, encapsulating all that so that it sits in a struct that can
  be accessed by DAl impls as well as the non-DAL default.

TBD: The main goal of the "MD_DEMO" DAL is to subvert our current default
  of opening a PUT-stream to storage at open-time, instead of at
  write-time.  This will allow creating zero-length files to be purely
  metadata operations.  It probably could just have no-ops everywhere.