DAL basics working for OBJ, and partially-working for NO_OP.
open_data()/close_data() are analogous to open_md()/close_md().
They allow DAL-related macros to be simpler.

We took a short-cut in the encapsulation of the DAL context.  There is
commented-out code in dal.c which provides init/destroy impls that
install a dynamically-allocated ObjectStream into the DAL context.
But much of MarFS still looks directly at the ObjectStream in the
FileHandle.  For instance, some marfs-ops expect underlying stream-ops
to have a side-effect on FileHandle.os.written.  So, instead of using
an internal OS in the DAL context, we just have open_data() install a
pointer to the file-handle's OS, for use by DAL implementations that
want it.

The OBJ DAL provides the old default object-store behavior, and this
will be installed as the default for Repo configurations that don't
specify a DAL.  All DAL-ops for OBJ just make straight-through calls
to the corresponding function in object_stream.c

The NO_OP DAL skips all interactions with objects, providing a way for
metadata benchmarks to eliminate the cost of object-store activity.
However, NO_OP also doesn't bother zeroing the buffers, so a read of
any object will always appear to have random contents.

WARNING: NO_OP creates metadata that refers to non-existent objids.
  Also, fuse 'rm' with a NO_OP DAL, crashes in some cases.  NO_OP
  should be regarded as a potentially-dangerous experimental option,
  for now.