Added support for DELETE, including via the DAL.  Refactored object-streams.
All "deletes" in marfs actually just move to the trash.  The only time
objects and trash files are actually deleted is in the Garbage Collector,
which was required to use custom low-level calls to delete objects and
files.  Now, we wanted to make the Garbage Collector do deletes through the
DAL, so that it would be agnostic about Multi-Compnent versus
object-storage back-ends.

Adding DELETE to object-streams is awkward, because the object-stream
functions do everything with only the ObjectStream component of a
file-handle.  Objects-streams are opened, then read/written, then closed.
Delete is the first case of what in the MDAL would be called a
"context-free" operation, working directly on a "path", without associated
state.

Did some refactoring of the way object-streams are initialized, so stream
inits previously done for open can be re-used for a delete operator.  GC
now uses fake_filehandle_for_delete(), and then the new delete_data() in
common.c.