Implements a new hashing scheme for determining the location of an object in the multi-component DAL
The previous hashing scheme was susceptible to a bad configuration
as described in issue #188. The new scheme uses the original hash function
to determine the pod, and generates three "random" hash functions (the
three functions are reproducible based on the hash of the object id). These
three functions are usede to determine the capacity unit, scatter dir, and
start-block for the object.