Fix issue 200.  Renaming files having non-null post.md_path was causing problems.
We now provide args in stat_xattrs() and str_2_post() that allow the
caller to suppress reloading the md_path, when parsing a POST xattr.

The fact the the MD path is wrong for a non-trash file is not a
problem, but loading xattrs was overwriting the correct MD path, used
to manipulate MD.

The buggy behavior is not seen in the production deployment, because
MD files are not getting a POST xattr containing an md_path.  This
seems more like the intended/desired behavior.  I haven't been able to
find a version of code that behaves that way for me.