Add the new TimingData to MarFS_FileHandle (see erasureUtils 198f6b9956b9)
This avoids the need to malloc a static blob, in order to hold ad hoc
timing-data collected inside libne, during low-level operations.  The
timing-data optionally maintained in ne_handle can not survive the
free of the ne_handle in ne_close(), which is also potentially
something we want to time.  Thus, we had previously malloc'ed a blob,
to hold a subset of this data, but we in turn depended on pftool to
free it, which leaks memory in the case of MarFS fuse.

erasureUtils now allows us to supercede the TimingData in ne_handle by
passing in a ptr to our own TimingData.  Thus, no need to malloc,
because our TimningData lives beyond ne_close().