add path to error messages for fuse errors
The error messages in the fuse component for failed FS operations
do not include the paths, but this makes it difficult to correlate
the error with the user operation that triggered it. In some cases
it may be possible to correlate it with a previous log message
containing the path, but it will be easier to just always log the
path in addition to the error message any time an error occurs.

For operations which take two paths, like link or rename, both
paths are reported because it can be ambiguous which path causes
the error.