fuse: initialize marfs config before mounting fuse
This commit changes the initialization of the marfs fuse
context to occur before calling fuse_main(). The reason for this
is to exit the fuse process before mounting the fuse filesystem
when there are fatal errors. Once fuse_main() is called, exiting
the daemon process leaves fuse mounted, causing ENOTCONN errors
when accessing the mountpoint. However, if there are fatal errors
in the configuration, there is no reason to even mount fuse at all.

This commit also changes the errors in marfs_fuse_init() to output
to stderr instead of using the LOG macro. This works because now
the function always runs in the foreground so sending output to
stderr works, and these errors should always be printed even if the
binary is not compiled with logging enabled.